moved MinesLeftType to types.ts
